debut routing
