+++ src/routes/Accueil.tsx
@@ -1,0 +1,10 @@
+const Accueil = () => {
+
+    // ACCUEIL AVEC CHOIX PRENOM
+    return (
+        <>
+        </>
+    )
+}
+
+export default Accueil
